% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_generation.R
\name{alpha_1}
\alias{alpha_1}
\title{Alpha generator using strategy 1 as per Hubin et. al.
TODO: This is just a placeholder.}
\usage{
alpha_1(feature)
}
\arguments{
\item{feature}{The feature to generate alphas for}
}
\description{
Alpha generator using strategy 1 as per Hubin et. al.
TODO: This is just a placeholder.
}
