% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_generation.R
\name{alpha_3}
\alias{alpha_3}
\title{Alpha generator using strategy 3 as per Hubin et. al.}
\usage{
alpha_3(feature, data, loglik)
}
\arguments{
\item{feature}{The feature to generate alphas for}

\item{data}{The dataset used}

\item{loglik}{log likelihood function to use}
}
\description{
Alpha generator using strategy 3 as per Hubin et. al.
}
