% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{arcsinh}
\alias{arcsinh}
\title{arcsinh transform}
\usage{
arcsinh(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
arcsinh(x)
}
\description{
arcsinh transform
}
\examples{
arcsinh(2)

}
