% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{cos.rad}
\alias{cos.rad}
\title{Cosine function for degrees}
\usage{
\method{cos}{rad}(x)
}
\arguments{
\item{x}{The vector of values in degrees}
}
\value{
The cosine of x
}
\description{
Cosine function for degrees
}
