% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{cos_deg}
\alias{cos_deg}
\title{Cosine function for degrees}
\usage{
cos_deg(x)
}
\arguments{
\item{x}{The vector of values in degrees}
}
\value{
The cosine of x
}
\description{
Cosine function for degrees
}
\examples{
cos_deg(0)

}
