% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{create.feature}
\alias{create.feature}
\title{Create method for "feature" class}
\usage{
create.feature(transform, features, alphas = NULL)
}
\arguments{
\item{transform}{A numeric vector denoting the alphas to use}

\item{features}{A list of features to include}
}
\description{
Create method for "feature" class
}
