% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{exp.dbl}
\alias{exp.dbl}
\title{Double exponential function}
\usage{
\method{exp}{dbl}(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
e^(-abs(x))
}
\description{
Double exponential function
}
