% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{exp_dbl}
\alias{exp_dbl}
\title{Double exponential function}
\usage{
exp_dbl(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
e^(-abs(x))
}
\description{
Double exponential function
}
\examples{
exp_dbl(2)

}
