% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{expi}
\alias{expi}
\title{Exponential function of absolute values}
\usage{
expi(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
e^(abs(x))
}
\description{
Exponential function of absolute values
}
