\docType{package}

\name{FBMS-package}
\alias{FBMS-package}
\alias{FBMS}

\title{
\packageTitle{FBMS}
}
\description{
\packageDescription{FBMS}
}
\author{
    \strong{Maintainer}: Jon Lachmann \email{jon@lachmann.nu}

    Authors:
    \itemize{
        \item Jon Lachmann \email{jon@lachmann.nu}
        \item Aliaksandr Hubin \email{aliaksah@math.uio.no}
    }

    Other contributors:
    \itemize{
        \item Florian Frommlet \email{florian.frommlet@meduniwien.ac.at} [contributor]
        \item Geir Storvik \email{geirs@math.uio.no} [contributor]
    }
}

\references{
    Lachmann, J., Storvik, G., Frommlet, F., & Hubin, A. (2022).
    A subsampling approach for Bayesian model selection.
    International Journal of Approximate Reasoning, 151, 33-63. Elsevier.

    Hubin, A., Storvik, G., & Frommlet, F. (2021).
    Flexible Bayesian Nonlinear Model Configuration.
    Journal of Artificial Intelligence Research, 72, 901-942.

    Hubin, A., Frommlet, F., & Storvik, G. (2021).
    Reversible Genetically Modified MJMCMC.
    Under review in EYSM 2021.

    Hubin, A., & Storvik, G. (2018).
    Mode jumping MCMC for Bayesian variable selection in GLMM.
    Computational Statistics & Data Analysis, 127, 281-297. Elsevier.
}
