% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{gaussian.loglik.bic.irlssgd}
\alias{gaussian.loglik.bic.irlssgd}
\title{Log likelihood function for gaussian regression with a prior p(m)=r*sum(total_width), using subsampling.}
\usage{
gaussian.loglik.bic.irlssgd(y, x, model, complex, params)
}
\arguments{
\item{y}{A vector containing the dependent variable}

\item{x}{The matrix containing the precalculated features}

\item{model}{The model to estimate as a logical vector}

\item{complex}{A list of complexity measures for the features}

\item{params}{A list of parameters for the log likelihood, supplied by the user}
}
\description{
Log likelihood function for gaussian regression with a prior p(m)=r*sum(total_width), using subsampling.
}
