% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{gaussian.loglik}
\alias{gaussian.loglik}
\title{Log likelihood function for gaussian regression with a prior p(m)=sum(total_width).}
\usage{
gaussian.loglik(y, x, model, complex)
}
\arguments{
\item{model}{The model as a logical vector to estimate}

\item{complex}{A list of complexity measures for the features}

\item{data}{Data to be used for the estimation}

\item{formula}{The formula to be used for estimation}
}
\description{
Log likelihood function for gaussian regression with a prior p(m)=sum(total_width).
}
