% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods2.R
\name{gaussian.loglik2.g}
\alias{gaussian.loglik2.g}
\title{Log likelihood function for linear regression using Zellners g-prior}
\usage{
gaussian.loglik2.g(y, x, model, complex, params = NULL)
}
\arguments{
\item{y}{A vector containing the dependent variable}

\item{x}{The matrix containing the precalculated features}

\item{model}{The model to estimate as a logical vector}

\item{complex}{A list of complexity measures for the features}

\item{params}{A list of parameters for the log likelihood, supplied by the user}
}
\value{
A list with the log marginal likelihood combined with the log prior (crit) and the posterior mode of the coefficients (coefs).
}
\description{
Log likelihood function for linear regression using Zellners g-prior
}
\examples{
gaussian.loglik2.g(rnorm(100), matrix(rnorm(100)), TRUE, list(oc=1))

}
