% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gen.probs.gmjmcmc}
\alias{gen.probs.gmjmcmc}
\title{Generate a probability list for GMJMCMC (Genetically Modified MJMCMC)}
\usage{
gen.probs.gmjmcmc(transforms)
}
\arguments{
\item{transforms}{A list of the transformations used (to get the count).}
}
\description{
Generate a probability list for GMJMCMC (Genetically Modified MJMCMC)
}
