\name{GMJMCMC-package}
\alias{GMJMCMC-package}
\alias{GMJMCMC}
\docType{package}
\title{
\packageTitle{GMJMCMC}
}
\description{
\packageDescription{GMJMCMC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GMJMCMC}
\packageIndices{GMJMCMC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{GMJMCMC}

Maintainer: \packageMaintainer{GMJMCMC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
