% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{gmjmcmc.iact}
\alias{gmjmcmc.iact}
\title{Integrated Auto-Correlation Time}
\usage{
gmjmcmc.iact(x)
}
\arguments{
\item{x}{A matrix where each row is a model}
}
\description{
Integrated Auto-Correlation Time
}
