% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{gmjmcmc.parallel}
\alias{gmjmcmc.parallel}
\title{Run multiple gmjmcmc runs in parallel returning a list of all results.}
\usage{
gmjmcmc.parallel(runs, cores = getOption("mc.cores", 2L), ...)
}
\arguments{
\item{runs}{The number of runs to run}

\item{cores}{The number of cores to run on}

\item{...}{Parameters to pass to gmjmcmc}
}
\value{
Results from multiple gmjmcmc runs
}
\description{
Run multiple gmjmcmc runs in parallel returning a list of all results.
}
