% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc.R
\name{gmjmcmc}
\alias{gmjmcmc}
\title{Main algorithm for GMJMCMC}
\usage{
gmjmcmc(data, loglik.pi, T, N, probs)
}
\arguments{
\item{data}{The data to use in the algorithm}

\item{loglik.pi}{The the (log) density to explore}

\item{T}{The number of population iterations}

\item{N}{The number of iterations per population (total iterations = T*N)}

\item{probs}{A list of the various probability vectors to use TODO: specification of this}
}
\description{
Main algorithm for GMJMCMC
}
