% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{gmjmcmc.totdens}
\alias{gmjmcmc.totdens}
\title{Total explored density}
\usage{
gmjmcmc.totdens(models)
}
\arguments{
\item{models}{A list of models with their criteria values (i.e. marginal likelihoods)}
}
\description{
Total explored density
}
