% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc.R
\name{gmjmcmc.transition}
\alias{gmjmcmc.transition}
\title{Subalgorithm for generating a new population of features in GMJMCMC}
\usage{
gmjmcmc.transition(
  S.t,
  F.0,
  data,
  loglik.alpha,
  marg.probs,
  transforms,
  probs,
  params
)
}
\arguments{
\item{S.t}{The current population of features}

\item{F.0}{The initial population of features, i.e. the bare covariates}

\item{data}{The data used in the model, here we use it to generate alphas for new features}

\item{loglik.alpha}{The log likelihood function to optimize the alphas for}

\item{marg.probs}{The marginal inclusion probabilities of the current features}

\item{transforms}{The nonlinear transformations available}

\item{probs}{A list of the various probability vectors to use}

\item{params}{A list of the various parameters for all the parts of the algorithm}
}
\value{
The updated population of features, that becomes S.t+1
}
\description{
Subalgorithm for generating a new population of features in GMJMCMC
}
