% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{linear.g.prior.loglik}
\alias{linear.g.prior.loglik}
\title{Log likelihood function for linear regression using Zellners g-prior}
\usage{
linear.g.prior.loglik(y, x, model, complex, params)
}
\arguments{
\item{model}{The model as a logical vector to estimate}

\item{complex}{A list of complexity measures for the features}

\item{data}{Data to be used for the estimation}

\item{formula}{The formula to be used for estimation}
}
\description{
Log likelihood function for linear regression using Zellners g-prior
}
