% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{log.prior}
\alias{log.prior}
\title{Log model prior function}
\usage{
\method{log}{prior}(params, complex)
}
\arguments{
\item{params}{list of passed parameters of the likelihood in GMJMCMC}

\item{complex}{list of complexity measures of the features included into the model}
}
\value{
A numeric with the log  model prior.
}
\description{
Log model prior function
}
\examples{
log.prior(params = list(r=2), complex = list(oc = 2))

}
