% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R, R/likelihoods2.R
\name{log_prior}
\alias{log_prior}
\title{Log model prior function}
\usage{
log_prior(mlpost_params, complex)

log_prior(mlpost_params, complex)
}
\arguments{
\item{mlpost_params}{list of passed parameters of the likelihood in GMJMCMC}

\item{complex}{list of complexity measures of the features included into the model}
}
\value{
A numeric with the log model prior.

A numeric with the log model prior.
}
\description{
Log model prior function

Log model prior function
}
\examples{
log_prior(mlpost_params = list(r=2), complex = list(oc = 2))

log_prior(paramsmlpost_params = list(r=2), complex = list(oc = 2))

}
