% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logistic.loglik.alpha}
\alias{logistic.loglik.alpha}
\title{Log likelihood function for logistic regression for alpha calculation
This function is just the bare likelihood function}
\usage{
logistic.loglik.alpha(a, data, mu_func)
}
\arguments{
\item{a}{A vector of the alphas to be used}

\item{data}{The data to be used for calculation}

\item{mu_func}{The function linking the mean to the covariates,
as a string with the alphas as a[i].}
}
\value{
A numeric with the log likelihood.
}
\description{
Log likelihood function for logistic regression for alpha calculation
This function is just the bare likelihood function
}
