% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logistic.loglik}
\alias{logistic.loglik}
\title{Log likelihood function for logistic regression with a prior p(m)=sum(total_width)}
\usage{
logistic.loglik(data, model, formula, complex)
}
\arguments{
\item{data}{Data to be used for the estimation}

\item{model}{The model as a logical vector to estimate}

\item{formula}{The formula to be used for estimation}

\item{complex}{A list of complexity measures for the features}
}
\description{
Log likelihood function for logistic regression with a prior p(m)=sum(total_width)
}
