% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc_support.R
\name{marginal.probs}
\alias{marginal.probs}
\title{Function for calculating marginal inclusion probabilities of features given a list of models}
\usage{
marginal.probs(models)
}
\arguments{
\item{models}{The list of models to use.}
}
\description{
Function for calculating marginal inclusion probabilities of features given a list of models
}
