% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc_support.R
\name{marginal.probs.renorm}
\alias{marginal.probs.renorm}
\title{Function for calculating feature importance through renormalized model estimates}
\usage{
marginal.probs.renorm(models, type = "features")
}
\arguments{
\item{models}{The models to use.}

\item{type}{Select which probabilities are of interest, features or models}
}
\description{
Function for calculating feature importance through renormalized model estimates
}
