% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjmcmc.R
\name{mjmcmc.loop}
\alias{mjmcmc.loop}
\title{The main loop for the MJMCMC (Mode Jumping MCMC) algorithm, used in both MJMCMC and GMJMCMC (Genetically Modified MJMCMC)}
\usage{
mjmcmc.loop(
  data,
  complex,
  loglik.pi,
  model.cur,
  N,
  probs,
  params,
  sub = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{The data to use}

\item{complex}{The complexity measures of the data}

\item{loglik.pi}{The (log) density to explore}

\item{model.cur}{The model to start the loop at}

\item{N}{The number of iterations to run for}

\item{probs}{A list of the various probability vectors to use}

\item{params}{A list of the various parameters for all the parts of the algorithm}

\item{sub}{An indicator that if the likelihood is inexact and should be improved each model visit (EXPERIMENTAL!)}

\item{verbose}{A logical denoting if messages should be printed}
}
\value{
A list containing the following elements:
\item{models}{All visited models.}
\item{accept}{Number of accepted proposals of the chain.}
\item{lo.models}{All models visited during local optimization.}
\item{best.crit}{The highest log marginal probability of the visited models.}
\item{marg.probs}{Marginal probabilities of the features.}
\item{model.probs}{Marginal probabilities of all of the visited models.}
\item{model.probs.idx}{Indices of unique visited models.}
}
\description{
The main loop for the MJMCMC (Mode Jumping MCMC) algorithm, used in both MJMCMC and GMJMCMC (Genetically Modified MJMCMC)
}
