% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{model.function}
\alias{model.function}
\title{Function to generate a function string for a model consisting of features}
\usage{
model.function(model, features, transforms, link)
}
\arguments{
\item{features}{A list of features to include}

\item{transform}{A numeric vector denoting the alphas to use}
}
\description{
Function to generate a function string for a model consisting of features
}
