% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{ngelu}
\alias{ngelu}
\title{Negative GELU function}
\usage{
ngelu(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
-x*pnorm(-x)
}
\description{
Negative GELU function
}
\examples{
ngelu(2)

}
