% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{nhs}
\alias{nhs}
\title{negative heavy side function}
\usage{
nhs(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
as.integer(x<0)
}
\description{
negative heavy side function
}
\examples{
nhs(2)

}
