% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{p0p05}
\alias{p0p05}
\title{p0p05 polynomial term}
\usage{
p0p05(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
p0(x)*(abs(x)+.Machine$double.eps)^(0.5)
}
\description{
p0p05 polynomial term
}
\examples{
p0p05(2)

}
