% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{p0pm05}
\alias{p0pm05}
\title{p0pm05 polynomial term}
\usage{
p0pm05(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
p0(x)\emph{sign(x)}(abs(x)+.Machine$double.eps)^(-0.5)
}
\description{
p0pm05 polynomial term
}
\examples{
p0pm05(2)

}
