% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{p0pm1}
\alias{p0pm1}
\title{p0pm1 polynomial terms}
\usage{
p0pm1(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
p0(x)*(x+.Machine$double.eps)^(-1)
}
\description{
p0pm1 polynomial terms
}
\examples{
p0pm1(2)

}
