% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{p0pm2}
\alias{p0pm2}
\title{p0pm2 polynomial term}
\usage{
p0pm2(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
p0(x)\emph{sign(x)}(abs(x)+.Machine$double.eps)^(-2)
}
\description{
p0pm2 polynomial term
}
\examples{
p0pm2(2)

}
