% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot.diagn}
\alias{plot.diagn}
\title{Plot convergence of best/median/mean/other summary log posteriors in time}
\usage{
\method{plot}{diagn}(res, FUN = median, conf = 0.95, burnin = 6, window = 5)
}
\arguments{
\item{res}{Object corresponding gmjmcmc output}

\item{FUN}{The summary statistics to check convergence}

\item{conf}{which confidence intervals to plot}

\item{burnin}{how many first populations to skip}

\item{window}{sliding window for computing the standard deviation}
}
\value{
summary statistics with given confidence intervals
}
\description{
Plot convergence of best/median/mean/other summary log posteriors in time
}
