% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.gmjmcmc_merged}
\alias{plot.gmjmcmc_merged}
\title{Plot a gmjmcmc_merged run}
\usage{
\method{plot}{gmjmcmc_merged}(x, count = "all", ...)
}
\arguments{
\item{x}{The results to use}

\item{count}{The number of features to plot, defaults to all}

\item{...}{Not used.}
}
\value{
No return value, just creates a plot
}
\description{
Plot a gmjmcmc_merged run
}
