% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.mjmcmc}
\alias{plot.mjmcmc}
\title{Function to plot the results, works both for results from gmjmcmc and
merged results from merge.results}
\usage{
\method{plot}{mjmcmc}(results, count = "all")
}
\arguments{
\item{results}{The results to use}

\item{count}{The number of features to plot, defaults to all}
}
\description{
Function to plot the results, works both for results from gmjmcmc and
merged results from merge.results
}
