% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{pm05}
\alias{pm05}
\title{pm05 polynomial term}
\usage{
pm05(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
(abs(x)+.Machine$double.eps)^(-0.5)
}
\description{
pm05 polynomial term
}
\examples{
pm05(2)

}
