% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{pm1}
\alias{pm1}
\title{pm1 polynomial term}
\usage{
pm1(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
sign(x)*(abs(x)+.Machine$double.eps)^(-1)
}
\description{
pm1 polynomial term
}
\examples{
pm1(2)

}
