% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{pm2}
\alias{pm2}
\title{pm2 polynomial term}
\usage{
pm2(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
sign(x)*(abs(x)+.Machine$double.eps)^(-2)
}
\description{
pm2 polynomial term
}
\examples{
pm2(2)

}
