% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gmjmcmc.2}
\alias{predict.gmjmcmc.2}
\title{New idea for a more streamlined function...
Produces slightly different results from the fun above since this is using all lo.models too.}
\usage{
\method{predict}{gmjmcmc.2}(
  object,
  x,
  link = function(x) x,
  quantiles = c(0.025, 0.5, 0.975),
  pop = 1
)
}
\description{
New idea for a more streamlined function...
Produces slightly different results from the fun above since this is using all lo.models too.
}
