% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gmjmcmc_merged}
\alias{predict.gmjmcmc_merged}
\title{Predict using a BGNLM model.}
\usage{
\method{predict}{gmjmcmc_merged}(object, x, link = function(x) x, quantiles = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{The model to use.}

\item{x}{The new data to use for the prediction, a matrix where each row is an observation.}

\item{link}{The link function to use}

\item{quantiles}{The quantiles to calculate credible intervals for the posterior moddes (in model space).}
}
\description{
Predict using a BGNLM model.
}
