% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{pseudo_log_1}
\alias{pseudo_log_1}
\title{Pseudo log transform}
\usage{
pseudo_log_1(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
asinh(x / (2 * sigma)) / log(base)
}
\description{
Pseudo log transform
}
\examples{
pseudo_log_1(2)

}
