% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{rmclapply}
\alias{rmclapply}
\title{rmclapply: Cross-platform mclapply/forking hack for Windows}
\usage{
rmclapply(runs, args, fun, mc.cores = NULL)
}
\arguments{
\item{runs}{The runs to run}

\item{args}{The arguments to pass to fun}

\item{fun}{The function to run}

\item{mc.cores}{Number of cores to use for parallel processing. Defaults to \code{detectCores()}.}
}
\value{
A list of results, with one element for each element of \code{X}.
}
\description{
This function applies a function in parallel to a list or vector (\code{X}) using multiple cores.
On Linux/macOS, it uses \code{mclapply}, while on Windows it uses a hackish version of parallelism.
The Windows version is based on \code{parLapply} to mimic forking following Nathan VanHoudnos.
}
