% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc_support.R
\name{set.transforms}
\alias{set.transforms}
\title{Set the transformations option for GMJMCMC (Genetically Modified MJMCMC),
this is also done when running the algorithm, but this function allows for it to be done manually.}
\usage{
set.transforms(transforms)
}
\arguments{
\item{transforms}{The vector of non-linear transformations}
}
\value{
No return value, just sets the gmjmcmc-transformations option
}
\description{
Set the transformations option for GMJMCMC (Genetically Modified MJMCMC),
this is also done when running the algorithm, but this function allows for it to be done manually.
}
\examples{
set.transforms(c("p0","p1"))


}
