% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
The sigmoid of x
}
\description{
Sigmoid function
}
\examples{
sigmoid(2)


}
