% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{sin_deg}
\alias{sin_deg}
\title{Sine function for degrees}
\usage{
sin_deg(x)
}
\arguments{
\item{x}{The vector of values in degrees}
}
\value{
The sine of x
}
\description{
Sine function for degrees
}
\examples{
sin_deg(0)

}
