% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{sqroot}
\alias{sqroot}
\title{Square root function}
\usage{
sqroot(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
The square root of the absolute value of x
}
\description{
Square root function
}
\examples{
sqroot(4)

}
