% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{string.population}
\alias{string.population}
\title{Function to get a character respresentation of a list of features
A list or a population of features in a character representation}
\usage{
string.population(x, round = 2)
}
\arguments{
\item{x}{A list of feature objects}

\item{round}{Rounding precision for parameters of the features}
}
\description{
Function to get a character respresentation of a list of features
A list or a population of features in a character representation
}
