% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.gmjmcmc}
\alias{summary.gmjmcmc}
\title{Function to print a quick summary of the results}
\usage{
\method{summary}{gmjmcmc}(object, pop = "last", tol = 1e-04)
}
\arguments{
\item{object}{The results to use}

\item{pop}{The population to print for, defaults to last}
}
\description{
Function to print a quick summary of the results
}
