% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.gmjmcmc_merged}
\alias{summary.gmjmcmc_merged}
\title{Function to print a quick summary of the results}
\usage{
\method{summary}{gmjmcmc_merged}(object, tol = 1e-04, labels = F, ...)
}
\arguments{
\item{object}{The results to use}

\item{tol}{The tolerance to use as a threshold when reporting the results.}

\item{labels}{Should the covariates be named, or just referred to as their place in the data.frame.}

\item{...}{Not used.}
}
\description{
Function to print a quick summary of the results
}
