% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.gmjmcmcresult}
\alias{summary.gmjmcmcresult}
\title{Function to print a quick summary of the results}
\usage{
\method{summary}{gmjmcmcresult}(results, pop = "last")
}
\arguments{
\item{results}{The results to use}

\item{pop}{The population to print for, defaults to last}
}
\description{
Function to print a quick summary of the results
}
