% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc_support.R
\name{summary.gmjresult}
\alias{summary.gmjresult}
\title{Summarize results from GMJMCMC}
\usage{
\method{summary}{gmjresult}(results, population = "last")
}
\arguments{
\item{results}{The results from GMJMCMC}

\item{populations}{A list of the populations to include in the summary, defaults to the last one}
}
\description{
Summarize results from GMJMCMC
}
