% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{to23}
\alias{to23}
\title{To the 2.3  power function}
\usage{
to23(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
x^2.3
}
\description{
To the 2.3  power function
}
\examples{
to23(2)

}
