% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_functions.R
\name{to72}
\alias{to72}
\title{To the 7/2  power function}
\usage{
to72(x)
}
\arguments{
\item{x}{The vector of values}
}
\value{
x^(7/2)
}
\description{
To the 7/2  power function
}
\examples{
to72(2)

}
