% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{update.alphas}
\alias{update.alphas}
\title{Update alphas on a feature}
\usage{
\method{update}{alphas}(feature, alphas, recurse = FALSE)
}
\arguments{
\item{feature}{The feature to be updated}

\item{alphas}{The alphas that will be used}

\item{recurse}{If we are recursing, to note the number of alphas used}
}
\description{
Update alphas on a feature
}
