% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{var.loglik}
\alias{var.loglik}
\title{Log likelihood function for a VAR model}
\usage{
var.loglik(y, x, model, complex, params)
}
\arguments{
\item{y}{A matrix containing the dependent variables}

\item{x}{The matrix containing the precalculated features}

\item{model}{The model to estimate as a logical vector}

\item{complex}{A list of complexity measures for the features}

\item{params}{A list of parameters for the log likelihood, supplied by the user}
}
\description{
Log likelihood function for a VAR model
}
