% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{weighted.quantiles}
\alias{weighted.quantiles}
\title{Calculate weighted quantiles}
\usage{
weighted.quantiles(values, weights, prob = c(0.025, 0.975))
}
\arguments{
\item{values}{The values to use}

\item{weights}{The weights of the values}

\item{prob}{The probabilities of the quantiles to use}
}
\value{
Weighted quantiles
}
\description{
Calculate weighted quantiles
}
