% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_generation.R
\name{alpha_2}
\alias{alpha_2}
\title{Alpha generator using strategy 1 as per Hubin et. al.}
\usage{
alpha_2(feature, transforms)
}
\arguments{
\item{feature}{The feature to generate alphas for}

\item{transforms}{The transforms used}
}
\description{
Alpha generator using strategy 1 as per Hubin et. al.
}
