% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gen.params.gmjmcmc}
\alias{gen.params.gmjmcmc}
\title{Generate a parameter list for GMJMCMC (Genetically Modified MJMCMC)}
\usage{
gen.params.gmjmcmc(data)
}
\arguments{
\item{data}{The dataset that will be used in the algorithm}
}
\value{
A list of parameters to use when running the mjmcmc function.
}
\description{
Generate a parameter list for GMJMCMC (Genetically Modified MJMCMC)
}
\examples{
gen.params.gmjmcmc(matrix(rnorm(600), 100))


}
