% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gen.params.list}
\alias{gen.params.list}
\title{Generate a parameter list for GMJMCMC}
\usage{
gen.params.list(data)
}
\arguments{
\item{data}{The dataset that will be used in the algorithm}
}
\description{
Generate a parameter list for GMJMCMC
}
