% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gen.params.mjmcmc}
\alias{gen.params.mjmcmc}
\title{Generate a parameter list for MJMCMC (Mode Jumping MCMC)}
\usage{
gen.params.mjmcmc(data)
}
\arguments{
\item{data}{The dataset that will be used in the algorithm}
}
\value{
A list of parameters to use when running the mjmcmc function.

Note that the $loglik item is an empty list, which is passed to the log likelihood function of the model,
intended to store parameters that the estimator function should use.
}
\description{
Generate a parameter list for MJMCMC (Mode Jumping MCMC)
}
\examples{
gen.params.mjmcmc(matrix(rnorm(600), 100))


}
