% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gen.probs.gmjmcmc}
\alias{gen.probs.gmjmcmc}
\title{Generate a probability list for GMJMCMC (Genetically Modified MJMCMC)}
\usage{
gen.probs.gmjmcmc(transforms)
}
\arguments{
\item{transforms}{A list of the transformations used (to get the count).}
}
\value{
A list of probabilities to be used as input for the gmjmcmc function.
}
\description{
Generate a probability list for GMJMCMC (Genetically Modified MJMCMC)
}
\examples{
gen.probs.gmjmcmc(c("p0", "exp_dbl"))


}
