% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gen.probs.list}
\alias{gen.probs.list}
\title{Generate a probability list for GMJMCMC}
\usage{
gen.probs.list(transforms)
}
\arguments{
\item{transforms}{A list of the transformations used (to get the count)}
}
\description{
Generate a probability list for GMJMCMC
}
