% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{gen.probs.mjmcmc}
\alias{gen.probs.mjmcmc}
\title{Generate a probability list for MJMCMC (Mode Jumping MCMC)}
\usage{
gen.probs.mjmcmc()
}
\value{
A list of probabilities to be used as input for the mjmcmc function.
}
\description{
Generate a probability list for MJMCMC (Mode Jumping MCMC)
}
\examples{
gen.probs.mjmcmc()

}
