\name{GMJMCMC-package}
\alias{GMJMCMC-package}
\alias{GMJMCMC}
\docType{package}
\title{
\packageTitle{GMJMCMC}
}
\description{
\packageDescription{GMJMCMC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GMJMCMC}
\packageIndices{GMJMCMC}

}
\author{
\packageAuthor{GMJMCMC}

Maintainer: \packageMaintainer{GMJMCMC}
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{
# simple examples of the most important functions
}
