% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{gmjmcmc.parallel}
\alias{gmjmcmc.parallel}
\title{Run multiple gmjmcmc (Genetically Modified MJMCMC) runs in parallel returning a list of all results.}
\usage{
gmjmcmc.parallel(
  x,
  y,
  loglik.pi = NULL,
  mlpost_params = list(family = "gaussian", beta_prior = list(type = "g-prior")),
  loglik.alpha = gaussian.loglik.alpha,
  transforms,
  runs = 2,
  cores = getOption("mc.cores", 2L),
  verbose = FALSE,
  merge.options = list(populations = "best", complex.measure = 2, tol = 1e-07),
  ...
)
}
\arguments{
\item{x}{matrix containing the design matrix with data to use in the algorithm}

\item{y}{response variable}

\item{loglik.pi}{The (log) density to explore}

\item{mlpost_params}{parameters for the estimator function loglik.pi}

\item{loglik.alpha}{The likelihood function to use for alpha calculation}

\item{transforms}{A Character vector including the names of the non-linear functions to be used by the modification}

\item{runs}{The number of runs to run}

\item{cores}{The number of cores to run on}

\item{verbose}{A logical denoting if messages should be printed}

\item{merge.options}{A list of options to pass to the \code{\link[=merge_results]{merge_results()}} function run after the run}

\item{...}{Further parameters passed to mjmcmc.}
}
\value{
Results from multiple gmjmcmc runs
}
\description{
Run multiple gmjmcmc (Genetically Modified MJMCMC) runs in parallel returning a list of all results.
}
\examples{
result <- gmjmcmc.parallel(
  runs = 1,
  cores = 1,
  loglik.pi = NULL,
  y = matrix(rnorm(100), 100),
  x = matrix(rnorm(600), 100),
  transforms = c("p0", "exp_dbl")
)

summary(result)

plot(result)

}
