% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc.R
\name{gmjmcmc}
\alias{gmjmcmc}
\title{Main algorithm for GMJMCMC}
\usage{
gmjmcmc(
  data,
  loglik.pi,
  loglik.alpha,
  transforms,
  T,
  N,
  N.final,
  probs,
  params,
  sub = F
)
}
\arguments{
\item{data}{A matrix containing the data to use in the algorithm,
first column should be the dependent variable, second should be the intercept
and the rest of the columns should be the independent variables.}

\item{loglik.pi}{The (log) density to explore}

\item{loglik.alpha}{The likelihood function to use for alpha calculation}

\item{transforms}{A list of the available nonlinear transformations for feature generation}

\item{T}{The number of population iterations}

\item{N}{The number of iterations per population (total iterations = (T-1)*N+N.final)}

\item{N.final}{The number of iterations for the final population (total iterations = (T-1)*N+N.final)}

\item{probs}{A list of the various probability vectors to use}

\item{params}{A list of the various parameters for all the parts of the algorithm}

\item{sub}{An indicator that if the likelihood is inexact and should be improved each model visit (EXPERIMENTAL!)}
}
\description{
Main algorithm for GMJMCMC
}
