% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc.R
\name{gmjmcmc}
\alias{gmjmcmc}
\title{Main algorithm for GMJMCMC}
\usage{
gmjmcmc(data, loglik.pi, transforms, T, N, N.final, probs, params)
}
\arguments{
\item{data}{The data to use in the algorithm}

\item{loglik.pi}{The (log) density to explore}

\item{transforms}{A list of the available nonlinear transformations for feature generation}

\item{T}{The number of population iterations}

\item{N}{The number of iterations per population (total iterations = T*N)}

\item{probs}{A list of the various probability vectors to use}

\item{params}{A list of the various parameters for all the parts of the algorithm}
}
\description{
Main algorithm for GMJMCMC
}
