% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{lm.logpost.bas}
\alias{lm.logpost.bas}
\title{Log likelihood function for Gaussian regression with parameter priors from BAS package
This function is created as an example of how to create an estimator that is used
to calculate the marginal likelihood of a model.}
\usage{
lm.logpost.bas(
  y,
  x,
  model,
  complex,
  params = list(r = exp(-0.5), prior_beta = "g-prior", alpha = 4)
)
}
\arguments{
\item{y}{A vector containing the dependent variable}

\item{x}{The matrix containing the precalculated features}

\item{model}{The model to estimate as a logical vector}

\item{complex}{A list of complexity measures for the features}

\item{params}{A list of parameters for the log likelihood, supplied by the user, important to specify the tuning parameters of beta priors and in Gaussian models}
}
\value{
A list with the log marginal likelihood combined with the log prior (crit) and the posterior mode of the coefficients (coefs).
}
\description{
Log likelihood function for Gaussian regression with parameter priors from BAS package
This function is created as an example of how to create an estimator that is used
to calculate the marginal likelihood of a model.
}
\examples{
lm.logpost.bas(rnorm(100), cbind(1,matrix(rnorm(100))), c(TRUE,TRUE), list(oc = 1))


}
