% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logistic.loglik}
\alias{logistic.loglik}
\title{Log likelihood function for logistic regression with a prior p(m)=sum(total_width)
This function is created as an example of how to create an estimator that is used
to calculate the marginal likelihood of a model.}
\usage{
logistic.loglik(y, x, model, complex, params)
}
\arguments{
\item{model}{The model as a logical vector to estimate}

\item{complex}{A list of complexity measures for the features}

\item{data}{Data to be used for the estimation}

\item{formula}{The formula to be used for estimation}
}
\description{
Log likelihood function for logistic regression with a prior p(m)=sum(total_width)
This function is created as an example of how to create an estimator that is used
to calculate the marginal likelihood of a model.
}
