% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmjmcmc.R
\name{mjmcmc.prop}
\alias{mjmcmc.prop}
\title{Subalgorithm for generating a proposal and acceptance probability}
\usage{
mjmcmc.prop(
  data,
  loglik.pi,
  model.cur,
  features,
  complex,
  marg.probs,
  probs,
  params
)
}
\arguments{
\item{data}{The data to use in the algorithm}

\item{loglik.pi}{The the (log) density to explore}

\item{model.cur}{The current model to make the proposal respective to}

\item{features}{The features available}

\item{probs}{A list of the various probability vectors to use TODO: specification of this}

\item{params}{A list of the various parameters for all the parts of the algorithm TODO: specification of this}
}
\description{
Subalgorithm for generating a proposal and acceptance probability
}
