% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjmcmc.R
\name{mjmcmc.prop}
\alias{mjmcmc.prop}
\title{Subalgorithm for generating a proposal and acceptance probability in (G)MJMCMC}
\usage{
mjmcmc.prop(
  data,
  loglik.pi,
  model.cur,
  complex,
  pip_estimate,
  probs,
  params,
  visited.models = NULL
)
}
\arguments{
\item{data}{The data to use in the algorithm}

\item{loglik.pi}{The the (log) density to explore}

\item{model.cur}{The current model to make the proposal respective to}

\item{complex}{The complexity measures used when evaluating the marginal likelihood}

\item{pip_estimate}{The current posterior inclusion probability estimate, used for proposals}

\item{probs}{A list of the various probability vectors to use}

\item{params}{A list of the various parameters for all the parts of the algorithm}

\item{visited.models}{A list of the previously visited models to use when subsampling and avoiding recalculation}
}
\description{
Subalgorithm for generating a proposal and acceptance probability in (G)MJMCMC
}
