% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjmcmc.R
\name{mjmcmc}
\alias{mjmcmc}
\title{Main algorithm for MJMCMC (Genetically Modified MJMCMC)}
\usage{
mjmcmc(data, loglik.pi, N = 100, probs = NULL, params = NULL, sub = FALSE)
}
\arguments{
\item{data}{A matrix containing the data to use in the algorithm,
first column should be the dependent variable, second should be the intercept
and the rest of the columns should be the independent variables.}

\item{loglik.pi}{The (log) density to explore}

\item{N}{The number of iterations to run for}

\item{probs}{A list of the various probability vectors to use}

\item{params}{A list of the various parameters for all the parts of the algorithm}

\item{sub}{An indicator that if the likelihood is inexact and should be improved each model visit (EXPERIMENTAL!)}
}
\description{
Main algorithm for MJMCMC (Genetically Modified MJMCMC)
}
