% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.gmjmcmc_merged}
\alias{plot.gmjmcmc_merged}
\title{Plot a gmjmcmc_merged run}
\usage{
\method{plot}{gmjmcmc_merged}(x, count = "all", pop = NULL, tol = 1e-07, data = NULL, ...)
}
\arguments{
\item{x}{The results to use}

\item{count}{The number of features to plot, defaults to all}

\item{pop}{The population to plot, defaults to last}

\item{tol}{The tolerance to use for the correlation when finding equivalent features, default is 0.0000001}

\item{data}{Data to merge on, important if pre-filtering was used}

\item{...}{Not used.}
}
\value{
No return value, just creates a plot
}
\description{
Plot a gmjmcmc_merged run
}
\examples{
result <- gmjmcmc.parallel(
 runs = 1,
 cores = 1,
 y = matrix(rnorm(100), 100),
 x = matrix(rnorm(600), 100),
 P = 2,
 transforms = c("p0", "exp_dbl")
)
plot(result)

}
