% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot.mjmcmc}
\alias{plot.mjmcmc}
\title{Function to plot the results, works both for results from gmjmcmc and
merged results from merge.results}
\usage{
\method{plot}{mjmcmc}(x, count = "all", ...)
}
\arguments{
\item{x}{The results to use}

\item{count}{The number of features to plot, defaults to all}

\item{...}{Not used.}
}
\description{
Function to plot the results, works both for results from gmjmcmc and
merged results from merge.results
}
