% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gmjmcmc.2}
\alias{predict.gmjmcmc.2}
\title{New idea for a more streamlined function...
Produces slightly different results from the fun above since this is using all lo.models too.}
\usage{
\method{predict}{gmjmcmc.2}(
  object,
  x,
  link = function(x) x,
  quantiles = c(0.025, 0.5, 0.975),
  pop = 1,
  ...
)
}
\arguments{
\item{object}{The model to use.}

\item{x}{The new data to use for the prediction, a matrix where each row is an observation.}

\item{link}{The link function to use}

\item{quantiles}{The quantiles to calculate credible intervals for the posterior moddes (in model space).}

\item{pop}{The population to use.}

\item{...}{Not used.}
}
\description{
New idea for a more streamlined function...
Produces slightly different results from the fun above since this is using all lo.models too.
}
