% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{print.feature}
\alias{print.feature}
\title{Print method for "feature" class}
\usage{
\method{print}{feature}(feature, transforms)
}
\arguments{
\item{x}{An object of class "feature"}
}
\description{
Print method for "feature" class
}
