% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{print.feature}
\alias{print.feature}
\title{Print method for "feature" class}
\usage{
\method{print}{feature}(feature, dataset = FALSE, alphas = FALSE, labels = FALSE, round = FALSE)
}
\arguments{
\item{feature}{An object of class "feature"}

\item{dataset}{Set the regular covariates as columns in a dataset}

\item{alphas}{Print a "?" instead of actual alphas to prepare the output for alpha estimation}

\item{labels}{Should the covariates be named, or just referred to as their place in the data.frame.}

\item{round}{Should numbers be rounded when printing? Default is FALSE, otherwise it can be set to the number of decimal places.}

\item{transforms}{The list of transforms to use}
}
\description{
Print method for "feature" class
}
