% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.mjmcmc}
\alias{summary.mjmcmc}
\title{Function to print a quick summary of the results}
\usage{
\method{summary}{mjmcmc}(results, tol = 1e-04)
}
\arguments{
\item{results}{The results to use}
}
\description{
Function to print a quick summary of the results
}
