% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{summary.mjmcmc_parallel}
\alias{summary.mjmcmc_parallel}
\title{Function to print a quick summary of the results}
\usage{
\method{summary}{mjmcmc_parallel}(object, tol = 1e-04, ...)
}
\arguments{
\item{object}{The results to use}

\item{tol}{The tolerance to use as a threshold when reporting the results.}
}
\description{
Function to print a quick summary of the results
}
